More db infrastructure work

+++ CodeProse.Shifter/domain/ScheduledShift.cs
@@ -5,10 +5,10 @@ namespace CodeProse.Shifter.domain
     public class ScheduledShift
     {
         public virtual Guid Id { get; set; }
-        public virtual int StartHour { get; set; }
-        public virtual int StartMinute { get; set; }
-        public virtual int EndHour { get; set; }
-        public virtual int EndMinute { get; set; }
+        public virtual long StartHour { get; set; }
+        public virtual long StartMinute { get; set; }
+        public virtual long EndHour { get; set; }
+        public virtual long EndMinute { get; set; }
         public virtual bool RepeatsOnMonday { get; set; }
         public virtual bool RepeatsOnTuesday { get; set; }
         public virtual bool RepeatsOnWednesday { get; set; }
